Fix --metrics-file argument.
