+++ config/default_dataset_metrics.py
@@ -2,8 +2,6 @@ from lsst.verify.tasks import ConfigPpdbLoader
 # Import these modules to ensure the metrics are registered
 import lsst.ap.association.metrics  # noqa: F401
 
-config.jobFileTemplate = "ap_verify.metricTask{id}.{dataId}.verify.json"
-
 ppdbConfigs = ["totalUnassociatedDiaObjects"]
 config.measurers = ppdbConfigs
